requirements.txt
discord.py==2.2.3
lyricsgenius==3.0.1
yt-dlp
PyNaCl==1.5.0
setuptools
